requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-cors-headers==3.0.2
django-extensions==2.2.1
django-filter==2.2.0
djangorestframework==3.10.1
idna==2.8
Markdown==3.1.1
pycodestyle==2.5.0
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
urllib3==1.25.3
